refactor: extract types, TemplateCard, and config forms from TemplateList.tsx
- TemplateListTypes.ts: form data types, constants, templateToForm
- TemplateCard.tsx: template card rendering
- TemplateConfigForms.tsx: AgentConfigForm, SkillConfigForm, ScriptConfigForm

TemplateList.tsx reduced from 807 to 275 lines.
